Major fix with the obj info not showing up
So the objects would get all of there values flushed and it would break
everything. well it wouldn't but it would clear the object that where
loaded before and only show the last one. i look around for ever because
i thought it wasn't writing right but it turns out it was just a for
loop that flushed the array was in the funtion before it was used. after
i moved it. Everything go fixed. so cool yeah wowoowowowowowowowowowoow